// src/types/enum/mp.ts
export enum ApplyEnterResultReasonEnum {
  PLAYER_JUDGE = 0,
  SCENE_CANNOT_ENTER = 1,
  PLAYER_CANNOT_ENTER_MP = 2,
  SYSTEM_JUDGE = 3,
  ALLOW_ENTER_PLAYER_FULL = 4,
  WORLD_LEVEL_LOWER_THAN_HOST = 5,
  HOST_IN_MATCH = 6,
  PLAYER_IN_BLACKLIST = 7,
  PS_PLAYER_NOT_ACCEPT_OTHERS = 8,
  HOST_IS_BLOCKED = 9,
  OTHER_DATA_VERSION_NOT_LATEST = 10,
  DATA_VERSION_NOT_LATEST = 11,
  PLAYER_NOT_IN_PLAYER_WORLD = 12,
  MAX_PLAYER = 13
}

export enum ClientReconnectReasonEnum {
  CLIENT_RECONNNECT_NONE = 0,
  CLIENT_RECONNNECT_QUIT_MP = 1
}

export enum MpSettingTypeEnum {
  MP_SETTING_NO_ENTER = 0,
  MP_SETTING_ENTER_FREELY = 1,
  MP_SETTING_ENTER_AFTER_APPLY = 2
}

export enum PreEnterMpStateEnum {
  INVALID = 0,
  START = 1,
  TIMEOUT = 2
}

export enum QuitReasonEnum {
  INVALID = 0,
  HOST_NO_OTHER_PLAYER = 1,
  KICK_BY_HOST = 2,
  BACK_TO_MY_WORLD = 3,
  KICK_BY_HOST_LOGOUT = 4,
  KICK_BY_HOST_BLOCK = 5,
  BE_BLOCKED = 6,
  KICK_BY_HOST_ENTER_HOME = 7,
  HOST_SCENE_INVALID = 8,
  KICK_BY_PLAY = 9
}